+++ party_calc/__init__.py
@@ -1,2 +1,2 @@
 __author__ = 'Boris Polyanskiy'
-__version__ = '1.1.2'
+__version__ = '1.1.3'
